+++ TableObjects/FoodCategoryObject.py
@@ -4,14 +4,15 @@ Food Category Object
 ---
 
 """
-class Food_Category:
+from .BaseTableObject import BaseTable
+
+class Food_Category(BaseTable):
     def __init__(self, id:int = -1, name:str = "Null"):
+        super().__init__(id)
         self.id = id
         self.name = name
     
-    """
-    Returns the SQLite command to create the object table.
-    """
+    
     @staticmethod
     def get_table_string() -> str:
         create_table_string: str = """
@@ -20,16 +21,16 @@ class Food_Category:
         name TEXT NOT NULL);"""
         return create_table_string
     
-    """
-    Converts the table into a SQLite row command
-    """
-    def to_sqlite(self):
-        if self.id < 0:
-            return f"""
+    def to_array(self) -> list:
+        return [self.id, self.name]
+        
+    def sqlite_insert(self) -> str:
+        return f"""
             INSERT INTO food_category (name)
             VALUES ("{self.name}");"""
-        else:
-            return f"""
+
+    def sqlite_update(self) -> str:
+        return f"""
             UPDATE food_category SET food_category_id = {self.id}, name = "{self.name}"
             WHERE food_category_id = {self.id};
             """