+++ TkinterOjects/__init__.py
@@ -1,0 +1,2 @@
+from .ChildFrameObject import ChildFrame
+from .ControllerMainObject import FrameController